+++ front_end/src/store/modules/cart/mutations.js
@@ -5,5 +5,4 @@ const mutations = {
   },
 };
 
-
 export default mutations;
