+++ front_end/src/store/modules/cart/state.js
@@ -1,0 +1,6 @@
+const state = {
+  cartItems: [],
+};
+
+
+export default state;
